Use common event loop interrupt on hardware error.
Just cleaning up the event info struct isn't enough: one some platforms,
there's a “drain loop” thread which runs underneath the monitor thread
to watch for an empty output buffer. This thread needs to be stopped,
too, or it'll segfault and bring down the whole JVM after the event info
struct is freed.