SerialImp.c:RXTXPort(nativeavailable) failure is OK on hidraw
With the previous two paches for handling multi-threading crashes,
the assumption for MacOS is hardware failure, when nativeavailable fails.

However for /dev/hidraw* devices `ioctl( fd, FIONREAD, &result )` fails.
It should set errno to ENOTTY, but sets it to EINVAL.

This change appears to be a hack.  When data is read after setting
RXTPPort.enableReceiveThreshold(>0) and
RXTXPort.enableReceiveTimeout(>0); nativeavaliable() is not called during
execution, read() expires on a timeout and it seems the eventloop is
not needed.  Optimal solution in such cases would be not to allocate
a thread for the event loop.